refactor(collapse): move into compute/
